h_snd.c: implement func_00109070 and func_0010a100 channel-clear logic
Both functions were guessed as calling the shared helpers
H_Snd_GetChannel/H_Snd_ClearChannel, which do not appear anywhere in
retail's compiled bytes for either function. Retail instead accesses
sChannels[index] directly (repeating the array-index computation
separately for each field touched, rather than caching a single
`channel` pointer) and inlines the clear sequence's three field writes
directly in a specific order (state, then active, then id) that does
not match H_Snd_ClearChannel's own declared order (active, state, id).

Both functions also share an odd but genuine retail idiom already
established elsewhere in this file (see the already-matching
func_00109e30): the channel-clear body is duplicated verbatim inside
an `if (channelIndex < 2) { ... } else { ... }` with byte-identical
logic on both arms, rather than being written once outside the branch
or shared via a helper call. This produces two separately-addressed
but identical compiled blocks; collapsing them into one shared branch
does not reproduce retail's bytes.

func_00109070: dropped the H_Snd_GetChannel call (retail has no `a0<0`
bound check, just `a0<2` gating which of the two identical arms to
take) and changed the parameter from s16 to s32 (retail's first
instruction operates on $a0 directly with no sign-extension prologue).
Result: MATCH (0 diff, 256B object against the 256B window).

func_0010a100: dropped H_Snd_GetChannel/H_Snd_ClearChannel/
H_Snd_ApplyChannelFade calls in favor of direct sChannels[] field
access and a direct func_0054d220(handle, fadeFrames) call (no
handle-non-NULL guard -- sChannels[i].active is already known true at
that point). Result: MATCH (0 diff, 312B object against the 320B
window).

Whole-file verify: 29 functions scanned, 23 MATCH (was 21), 6
NONMATCHING, no regressions.

+++ include/h_snd.h
@@ -157,7 +157,7 @@ s16 H_Snd_GetCurrentBgmId(void);
 void H_Snd_StopBgm(void);
 void H_Snd_StopBgmFade(s16 fadeDuration);
 u8 func_00109040(s16 id, s16 unused);
-u8 func_00109070(s16 channelIndex);
+u8 func_00109070(s32 channelIndex);
 u8 func_00109170(void);
 void H_Snd_00109180(s32 channelIndex);
 s16 func_001099a0(void);
